chore(maintner): renamed directories
Key changes include cmd => maintnerd
                    api => maintnerd/api
                    internal => maintnerd/internal

Other standalone utility tools are moved to the utils directory
